pyproject.toml
[project]
name = "hyperpytext"
version = "0.1.0"
description = "A tool to create web applications"
readme = "README.md"
authors = [{ name = "minollisantiago", email = "minollisantiago@gmail.com" }]
requires-python = ">=3.13"
dependencies = ["typer>=0.9.0", "rich>=13.7.0", "pyyaml>=6.0"]

[project.scripts]
create-hyperpy-app = "hyperpytext:app"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/hyperpytext"]

[tool.hatch.build]
include = ["src/hyperpytext/templates/**/*.yaml"]
